networksettings: Don't crash when the sim is ejected
If the current subscription ceases to exist, for instance, after
the SIM card is ejected, this preference screen would crash
when setting up the view.

This patch contains two parts, the first doesn't add non-existing
(null) subscriptions to the valid subscription list, the second
finishes the screen when the list of subscriptions is empty because
the screen is effectively useless.

Ticket: CYNGNOS-1794
Change-Id: Iedd349dc3757c62ad74e47927fc77109ea4ba8e6