[update] Allow to disable the new scan API for manual network search.
The implementation to attempt new API and then fallback after
error is generic but may be not efficient. For the devices that
is not capable of using new API, it lengthens the network search
because of the new API scan trial in the first place and may fail
if the error is not properly caught.
(Refer to Ifebbac9965a40acaff0d50d32ca8603c72a6a77f for details)

As an alternative, this commit provides a config for devices that
are known as not capable of using new API to disable new API scan
and start network search through old API directly.

[AICP]

Follow up to https://gerrit.aicp-rom.com/c/AICP/packages_services_Telephony/+/87887
Updated to latest patchset

Change-Id: Ibfbee447f49f5e9134b6191df23c96326c0c3521